requirements.txt
celery==5.5.3
fastapi==0.115.12
jose==1.0.0
passlib==1.7.4
Pillow==11.2.1
protobuf==4.25.3
pydantic==1.10.22
python-dotenv==1.1.0
redis==6.2.0
SQLAlchemy==2.0.34
uvicorn==0.24.0
psycopg2-binary==2.9.9
python-multipart==0.0.6
alembic==1.12.1
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
google-generativeai==0.3.2
